requirements.txt
webapp2
gcloud==0.18.3
google-api-python-client==1.4.2
google-apitools==0.4.11
httplib2==0.9.1
oauth2client==1.5.2
protobuf==3.5.1
protorpc==0.11.1
pyasn1==0.1.8
pyasn1-modules==0.0.7
pycrypto==2.6.1
rsa==3.2
simplejson==3.8.0
six==1.9.0
uritemplate==0.6